+++ src/Statistics.h
@@ -20,19 +20,19 @@ public:
     void IncrementCounter(std::string error, uint8_t fecId, uint64_t increment=1);
     long GetCounter(std::string error, uint8_t fecId);
    
-    double GetDeltaTriggerTimestamp(uint8_t fecId);
-    void SetDeltaTriggerTimestamp(uint8_t fecId, double val);
-    double GetOldTriggerTimestamp(uint8_t fecId);
-    void SetOldTriggerTimestamp(uint8_t fecId, double srsTimestamp);
-    double GetFirstTriggerTimestamp(uint8_t fecId);
-    void SetFirstTriggerTimestamp(uint8_t fecId, double srsTimestamp);
-    double GetMaxTriggerTimestamp(uint8_t fecId);
-    void SetMaxTriggerTimestamp(uint8_t fecId, double srsTimestamp);
+    uint64_t GetDeltaTriggerTimestamp(uint8_t fecId);
+    void SetDeltaTriggerTimestamp(uint8_t fecId, uint64_t val);
+    uint64_t GetOldTriggerTimestamp(uint8_t fecId);
+    void SetOldTriggerTimestamp(uint8_t fecId, uint64_t srsTimestamp);
+    uint64_t GetFirstTriggerTimestamp(uint8_t fecId);
+    void SetFirstTriggerTimestamp(uint8_t fecId, uint64_t srsTimestamp);
+    uint64_t GetMaxTriggerTimestamp(uint8_t fecId);
+    void SetMaxTriggerTimestamp(uint8_t fecId, uint64_t srsTimestamp);
     
-    double GetLowestCommonTriggerTimestampDet(uint8_t det);
-    void SetLowestCommonTriggerTimestampDet(uint8_t det, double val);
-    double GetLowestCommonTriggerTimestampPlane(std::pair<uint8_t, uint8_t> dp);
-    void SetLowestCommonTriggerTimestampPlane(std::pair<uint8_t, uint8_t> dp, double val);
+    uint64_t GetLowestCommonTriggerTimestampDet(uint8_t det);
+    void SetLowestCommonTriggerTimestampDet(uint8_t det, uint64_t val);
+    uint64_t GetLowestCommonTriggerTimestampPlane(std::pair<uint8_t, uint8_t> dp);
+    void SetLowestCommonTriggerTimestampPlane(std::pair<uint8_t, uint8_t> dp, uint64_t val);
    
     void PrintClusterStats(Configuration& config);
     void PrintFECStats(Configuration& config);
